requirements.txt
python-dotenv
faiss-cpu
google-genai
google-adk
streamlit
langchain
langchain-community
langchain-google-genai
faiss-cpu
sentence-transformers